mcp-service/sql_generator/requirements.txt
jinja2>=3.0.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
